pyproject.toml
[project]
name = "sequrity-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.96.0",
    "click>=8.3.1",
    "google-genai>=1.67.0",
    "ipykernel>=7.2.0",
    "openai>=2.29.0",
    "pre-commit>=4.5.1",
    "pygithub>=2.9.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rich>=14.3.3",
    "tenacity>=9.1.4",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
]

[tool.ruff]
line-length = 125
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I"]  # pycodestyle, pyflakes, isort
